Create a proper boot context class
